[b-7] Currency convertion.

+++ src/main/java/com/ylysenkova/movieland/model/Movie.java
@@ -11,7 +11,7 @@ public class Movie {
     private int yearOfRelease;
     private String description;
     private double rating;
-    private BigDecimal price;
+    private double price;
     private String picturePath;
     private List<Country> countries;
     private List<Genre> genres;
@@ -70,11 +70,11 @@ public class Movie {
         this.rating = rating;
     }
 
-    public BigDecimal getPrice() {
+    public double getPrice() {
         return price;
     }
 
-    public void setPrice(BigDecimal price) {
+    public void setPrice(double price) {
         this.price = price;
     }
